refactor(core): move TermFilter and TermFilterSupport to request.search.
...package
