+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/request/search/TermFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request;
+package com.b2international.snowowl.core.request.search;
 
 import java.io.Serializable;
