+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/request/search/TermFilterSupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2021 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2020-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request;
+package com.b2international.snowowl.core.request.search;
+
+import com.b2international.snowowl.core.request.SearchResourceRequest;
 
 /**
  * Interface to configure term filters in a given request
